Fix loop counter overflow due to inconsistent type
The modified loops used `int` counters but `uint32_t`, `size_t`, or `uint64_t`
upper bounds. This could have led to potential overflows during loop counter
increment when the upper bound exceeds (2^31)-1.

Additionally, since signed overflows are undefined as per the ISO C standard,
the overflowed loop counters could have subsequently triggered a buffer overflow
when used as array indices.

We fix these issues by using consistent types for both the loop counter and the
upper bound.

+++ stuffer/s2n_stuffer.c
@@ -326,7 +326,7 @@ int s2n_stuffer_writev_bytes(struct s2n_stuffer *stuffer, const struct iovec* io
     POSIX_ENSURE(S2N_MEM_IS_READABLE(ptr, size), S2N_ERR_NULL);
 
     size_t size_left = size, to_skip = offs;
-    for (int i = 0; i < iov_count; i++) {
+    for (size_t i = 0; i < iov_count; i++) {
         if (to_skip >= iov[i].iov_len) {
             to_skip -= iov[i].iov_len;
             continue;
